refactor: move session renew helpers
